+++ C#/Controllers/StaffController.cs
@@ -95,20 +95,21 @@ namespace CMSAPI.Controllers
     [Route("DeleteStaff")]
     public async Task<IActionResult> DeleteStaff(int id)
     {
-      try
-      {
-        var staff = await StaffRepository.DeleteStaff(id);
-        if (staff == null)
-        {
-          return NotFound();
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    await StaffRepository.DeleteStaff(id);
+                    return Ok();
+                }
+                catch (Exception)
+                {
+                    return BadRequest();
+                }
+
+            }
+            return BadRequest();
         }
-        return Ok(staff);
-      }
-      catch (Exception)
-      {
-        return BadRequest();
-      }
-    }
     #endregion
 
     #region GetStaffById
@@ -130,27 +131,6 @@ namespace CMSAPI.Controllers
         return BadRequest();
       }
     }
-        #endregion
-
-        #region Update Is Active
-        [HttpGet]
-        [Route("UpdateIsActive")]
-        public async Task<IActionResult> UpdateIsActive(int id)
-        {
-            try
-            {
-                var staff = await StaffRepository.UpdateIsActive(id);
-                if (staff == null)
-                {
-                    return NotFound();
-                }
-                return Ok(staff);
-            }
-            catch (Exception)
-            {
-                return BadRequest();
-            }
-        }
         #endregion
 
     }
